refactor: Rename FeedlyController to FeedlyClient
Renamed the class and corresponding file from FeedlyController to FeedlyClient
to better reflect its responsibility as an API client. Updated all imports.
