+++ feedly_regexp_marker/classifier.py
@@ -10,7 +10,7 @@ from typing import Literal, Mapping, Optional, cast
 
 from pydantic import BaseModel, ConfigDict, RootModel
 
-from feedly_regexp_marker.feedly_controller import Action, Entry, StreamId
+from feedly_regexp_marker.feedly_client import Action, Entry, StreamId
 from feedly_regexp_marker.pattern_texts import PatternTexts
 from feedly_regexp_marker.rules import Rule, Rules
