+++ feedly_regexp_marker/rules.py
@@ -6,7 +6,7 @@ from typing import Optional
 from pydantic import BaseModel, ConfigDict, RootModel
 from pydantic_yaml import parse_yaml_file_as
 
-from feedly_regexp_marker.feedly_controller import Action, StreamId
+from feedly_regexp_marker.feedly_client import Action, StreamId
 from feedly_regexp_marker.pattern_texts import PatternTexts
